day3/requirements.txt
langchain-core==0.1.28
langchain-openai==0.1.16
langchain-community==0.0.29
langchain-text-splitters==0.0.1
langgraph==0.0.36
tavily-python==0.3.2
python-dotenv==1.0.1
chroma-hnswlib==0.7.3
chromadb==0.4.24
tiktoken==0.5.2
typing-extensions==4.9.0 